+++ packages/eyes-api/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- make output classes readonly
+- move TestAccessibilityStatus to its own file
 - add variantId to CheckSettings
 
 ## 0.0.2 - 2021/3/24